build: add test coverage to frontend
Besides the advantage of having coverage information, this is also useful to make sure the frontend tests are rerun after a clean. Indeed, since the test task had no output, tests were not rerun after a clean if the test task inputs had not changed, which is a bit surprising.