feat(users): add CRUD services for users, and password resets/changes
The user model is limited to login for now, but it will probably contain roles and maybe email in the future.

The intended usage of the API is that an admin creates a user, gets the automatically generated password and gives the login and password to the actual user, which is then free to change his password.

We could imagine storing a flag indicating that the password is auto-generated and force the user to change it, but it's probably useless, since the admin can reset the password anyway, and thus act as the other users.