chore(tests): drastically speed tests, from around 48 seconds to around 9 seconds
the magic is to stop loading the AppModule, and, when possible, the RouterTestingModule, and to instead declare and provide only the dependencies that are necessary for the tests. The nice thing is that the more the app grew, the more time each test that loaded the AppModule took, making the performance worse and worse. This should keep things sane.