refactor: refactor backend authentication as a filter
See https://github.com/spring-projects/spring-boot/issues/11234#issuecomment-358069134: the Spring team seems to favor the usage of filters over handler interceptors. The downside is that we can't throw exceptions to send back error responses (they're transformed to 500 errors). The upside is that actuator requests can be handled the same way as other requests.