refactor: use a StorageProperties class to load the GCS properties
This is cleaner than using @Value.
Note that even though the properties must be in lower case with dashes
when reading them, then can be written in camelCase in the yaml file
or in the command-line parameters or environment variables.

Also note that, to be cleaner, I changed the name of the properties:

`globe42.googleCloudStorageCredentials` becomes
`globe42.googleCloudStorage.credentials`, and
`globe42.googleCloudStorageCredentialsPath` becomes
`globe42.googleCloudStorage.credentialsPath`.

In order to avoid forgetting to change the environment when deploying
this new version, I already added the new environment variable in the
Clever Cloud console. We'll just have to remove the old variable when
this change is deployed.