refactor: improve confirm service
Instead of emitting an error on close, which forces every caller to use an error callback in order to avoid an error in tests and in the browser console, the observable simply completes without emitting any event.
This makes the nominal case simpler.
If the caller wants to be notified when closed (like in a CanDeactivate guard), then the caller can pass `errorOnClose: true` in the confirm options.